// src/types/operators.ts
export enum LabelOperators {
  equal = 'равно',
  greater = 'больше',
  greater_or_equal = 'больше или равно',
  less = 'меньше',
  less_or_equal = 'меньше или равно',
  not_equal = 'не равно',
  begins_with = 'начинается с',
  not_begins_with = 'не начинается с',
  contains = 'содержит',
  not_contains = 'не содержит',
  in = 'из указанных',
  not_in = 'не из указанных',
  is_empty = 'пустая строка',
  is_not_empty = 'не пустая строка',
  is_null = 'пусто',
  is_not_null = 'не пусто',
  regexp = 'regexp',
  between = 'между',
}

export enum Operators {
  equal = 'equal',
  greater = 'greater',
  greater_or_equal = 'greater_or_equal',
  less = 'less',
  less_or_equal = 'less_or_equal',
  not_equal = 'not_equal',
  begins_with = 'begins_with',
  not_begins_with = 'not_begins_with',
  contains = 'contains',
  not_contains = 'not_contains',
  in = 'in',
  not_in = 'not_in',
  is_empty = 'is_empty',
  is_not_empty = 'is_not_empty',
  is_null = 'is_null',
  is_not_null = 'is_not_null',
  regexp = 'regexp',
  between = 'between',
}